Refactor the code with reviewers loading dart
